requirements.txt
opencv-python==4.12.0.88
easyocr==1.7.2
openai==2.1.0
python-dotenv==1.1.1
numpy==2.2.6
torch==2.8.0
torchvision==0.23.0
scipy==1.16.2
Pillow==11.3.0
scikit-image==0.25.2
python-bidi==0.6.6
PyYAML==6.0.3
Shapely==2.1.2
pyclipper==1.3.0.post6
ninja==1.13.0
anyio==4.11.0
distro==1.9.0
httpx==0.28.1
jiter==0.11.0
pydantic==2.11.9
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.15.0
